Add additional parameters, remove CSV parsing
