+++ agent/tool.py
@@ -1,5 +1,7 @@
 from enum import Enum
 import logging
+import os
+from pathlib import Path
 from typing import Type, Optional, Literal
 from string import Template
 from pandas import DataFrame
@@ -10,14 +12,51 @@ from langchain_core.callbacks import CallbackManagerForToolRun
 from langchain_core.pydantic_v1 import BaseModel, Field
 from langchain_core.tools import BaseTool
 
-# Debug
-from langchain.globals import set_verbose, set_debug
 
-from agent.listingscraper.parse import (
-    expand_to_csv_row,
-    format_human_readable,
-    save_row_data,
+
+human_readable_row_tpl = Template(
+    """
+---
+$mls_id: $street, $city, $zip_code
+
+property style: $style
+street: $street
+city: $city
+zip Code: $zip_code
+bedrooms: $beds bedrooms
+stories: $stories stories
+full baths: $full_baths
+half baths: $half_baths
+sqft: $sqft sqft
+listed price: $list_price
+sold for: $sold_price
+nearby schools: $nearby_schools
+listing url: $property_url
+
+description: $text
+---
+"""
 )
+
+
+def format_human_readable(property: DataFrame) -> str:
+    return human_readable_row_tpl.substitute(property.to_dict())
+
+
+def save_rows_to_csv(properties: DataFrame, filename: str = "listings.csv") -> str:
+    print(f"Saving rows data to {filename}...")
+
+    # todo: update path access to be cleaner
+    data_base_dir_path = Path(os.getcwd(), "data")
+    listings_csv_path = Path(data_base_dir_path, filename)
+
+    if not data_base_dir_path.exists():
+        data_base_dir_path.mkdir()
+
+    if not listings_csv_path.exists():
+        listings_csv_path.touch()
+
+    properties.to_csv(listings_csv_path)
 
 
 class ListingType(Enum):
@@ -34,7 +73,25 @@ class HomeSearchResultsInput(BaseModel):
     listing_type: ListingType = Field(
         default=ListingType.FOR_SALE, description="type of listing"
     )
-    radius: Optional[float] = None
+    radius: Optional[float] = Field(
+        description="""
+        Radius in miles to find comparable properties based on individual addresses. Example: 5.5 (fetches properties within a 5.5-mile radius if location is set."""
+    )
+
+    # Everything below is a param to the Tool, but not to the homeharvest scraper
+    bedroom_number: Optional[float] = Field(
+        description="""The number of bedrooms a user is looking for in a property. If not provided, it defaults to 2.0.
+    """
+    )
+    bathroom_number: Optional[float] = Field(
+        description="""The number of bathrooms a user is looking for in a property. If not provided, it defaults to 2.0."""
+    )
+    min_price: Optional[int] = Field(
+        description="""The minimum price of a property to search for. If not provided, it defaults to 10000000"""
+    )
+    max_price: Optional[int] = Field(
+        description="""The maximum price of a property to search for. If not provided, it defaults to 100000000"""
+    )
 
 
 # This subclasses langchain's BaseTool to create a custom Tool to pass into OpenAI.
@@ -135,10 +192,11 @@ class HomeSearchResultsTool(BaseTool):
             logging.info("Inferred bathroom_number for filtering: %s", bathroom_number)
             logging.info("Inferred listing_type for filtering: %s", listing_type)
             logging.info("Inferred radius for radius: %s", radius)
+            
+            save_rows_to_csv(properties)
 
             properties_expanded = []
             for _, row in properties.iloc[0 : self.max_results].iterrows():
-                # save_row_data(row)
                 expanded_row = format_human_readable(row)
                 properties_expanded.append(expanded_row)
         except Exception as e:
